backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-multipart
google-generativeai
python-dotenv
slowapi

